Rewiev and decomposition commands pool and identity
