+++ commands/identity/subcommands/TopUpIdentityCommand.js
@@ -1,9 +1,9 @@
-import BaseCommand from './BaseCommand.js';
-import topUpIdentityAction from "../actions/topUpIdentityAction.js";
+import BaseCommand from '../../BaseCommand.js';
+import topUpIdentityAction from "../../../actions/topUpIdentityAction.js";
 
 class TopUpIdentityCommand extends BaseCommand {
   constructor() {
-    super('topUpIdentity');
+    super('topup');
     this.description("Make TopUp Identity balance")
       .argument('<amount>', 'Amount credits for TopUp Identity balance')
       .action(topUpIdentityAction());
