+++ commands/pool/subcommand/CreatePoolCommand.js
@@ -1,10 +1,10 @@
-import BaseCommand from './BaseCommand.js';
-import createPoolAction from '../actions/createPoolAction.js'
-import MasternodeTypeEnum from "../models/enums/MasternodeTypeEnum.js";
+import BaseCommand from "../../BaseCommand.js";
+import createPoolAction from "../../../actions/createPoolAction.js";
+import MasternodeTypeEnum from "../../../models/enums/MasternodeTypeEnum.js";
 
 class CreatePoolCommand extends BaseCommand {
   constructor() {
-    super("createPool");
+    super("create");
     this.description("Create Pool")
       .argument('<name>', 'The name of the pool')
       .argument('<description>', 'The description of the pool')
